+++ src/anyvar/storage/mapper_registry.py
@@ -4,9 +4,8 @@ from typing import TypeVar
 
 from ga4gh.vrs import models as vrs_models
 
-from anyvar.storage import db
-
-from .mappers import (
+from anyvar.storage import orm
+from anyvar.storage.mappers import (
     AlleleMapper,
     BaseMapper,
     SequenceLocationMapper,
@@ -22,9 +21,9 @@ class MapperRegistry:
     def __init__(self) -> None:
         """Initialize the MapperRegistry with known mappers."""
         self._mappers: dict[type, BaseMapper] = {
-            db.Allele: AlleleMapper(),
-            db.Location: SequenceLocationMapper(),
-            db.SequenceReference: SequenceReferenceMapper(),
+            orm.Allele: AlleleMapper(),
+            orm.Location: SequenceLocationMapper(),
+            orm.SequenceReference: SequenceReferenceMapper(),
         }
 
     def get_mapper(self, entity_type: type[T]) -> BaseMapper:
@@ -43,9 +42,9 @@ class MapperRegistry:
         """Convert any VRS model to its corresponding DB entity."""
         # Map VRS model types to DB entity types
         vrs_to_db_mapping = {
-            vrs_models.Allele: db.Allele,
-            vrs_models.SequenceLocation: db.Location,
-            vrs_models.SequenceReference: db.SequenceReference,
+            vrs_models.Allele: orm.Allele,
+            vrs_models.SequenceLocation: orm.Location,
+            vrs_models.SequenceReference: orm.SequenceReference,
         }
 
         db_type = vrs_to_db_mapping.get(type(vrs_model))
